final-version/pyproject.toml
[project]
name = "projects"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.5",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "apscheduler==3.11.2",
    "astroid==3.1.0",
    "authlib==1.6.9",
    "beautifulsoup4==4.14.3",
    "blinker==1.9.0",
    "boto3==1.40.61",
    "botocore==1.40.61",
    "cachetools==6.2.6",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "coverage==7.13.4",
    "coze-coding-dev-sdk==0.5.11",
    "coze-coding-utils==0.2.4",
    "coze-workload-identity==0.1.4",
    "cozeloop==0.1.25",
    "cryptography==46.0.5",
    "cssselect==1.4.0",
    "cssutils==2.11.1",
    "dbus-python==1.3.2",
    "deprecation==2.1.0",
    "dill==0.4.1",
    "distro==1.9.0",
    "docx2python==3.5.0",
    "et-xmlfile==2.0.0",
    "fastapi==0.121.2",
    "flask==3.1.3",
    "fsspec==2026.2.0",
    "git-filter-repo==2.47.0",
    "gitdb==4.0.12",
    "gitignore-parser==0.1.13",
    "gitpython==3.1.45",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-ws==0.8.2",
    "hyperframe==6.1.0",
    "idna==3.11",
    "inflect==7.5.0",
    "iniconfig==2.3.0",
    "isort==5.13.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "jmespath==1.1.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.0.3",
    "langchain-core==1.0.2",
    "langchain-openai==1.0.1",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.0",
    "langgraph-checkpoint-postgres==3.0.1",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.39",
    "lxml==6.0.2",
    "mako==1.3.10",
    "markdown==3.10.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "more-itertools==10.8.0",
    "multidict==6.7.1",
    "numpy==2.2.6",
    "openai==2.24.0",
    "opencv-python==4.12.0.88",
    "openpyxl==3.1.5",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==25.0",
    "pandas==2.2.2",
    "paragraphs==1.0.1",
    "pdfminer-six==20251230",
    "pdfplumber==0.11.9",
    "pillow==10.3.0",
    "platformdirs==4.9.2",
    "pluggy==1.6.0",
    "postgrest==2.27.3",
    "propcache==0.4.1",
    "psutil==7.1.3",
    "psycopg==3.3.0",
    "psycopg-binary==3.3.0",
    "psycopg-pool==3.3.0",
    "psycopg2-binary==2.9.9",
    "pycparser==3.0",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pygments==2.19.2",
    "pygobject==3.48.2",
    "pyiceberg==0.11.1",
    "pyjwt==2.10.1",
    "pylint==3.1.0",
    "pymupdf==1.27.1",
    "pyparsing==3.3.2",
    "pypdf==6.4.1",
    "pypdfium2==5.6.0",
    "pyroaring==1.0.3",
    "pytest==9.0.1",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-pptx==1.0.2",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "realtime==2.27.3",
    "regex==2026.2.28",
    "reportlab==4.4.10",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "s3transfer==0.14.0",
    "setuptools==68.1.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "sqlacodegen==4.0.2",
    "sqlalchemy==2.0.44",
    "starlette==0.49.3",
    "storage3==2.27.3",
    "strenum==0.4.15",
    "strictyaml==1.7.3",
    "supabase==2.27.3",
    "supabase-auth==2.27.3",
    "supabase-functions==2.27.3",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tomlkit==0.14.0",
    "tqdm==4.67.3",
    "typeguard==4.5.1",
    "types-html5lib==1.1.11.20251117",
    "types-lxml==2026.2.16",
    "types-webencodings==0.5.0.20251108",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "unidecode==1.4.0",
    "urllib3==2.6.3",
    "uvicorn==0.38.0",
    "watchdog==6.0.0",
    "websockets==15.0.1",
    "werkzeug==3.1.6",
    "wheel==0.42.0",
    "wsproto==1.3.2",
    "xlrd==2.0.2",
    "xlsxwriter==3.2.9",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zstandard==0.25.0",
]

[tool.uv]
# 使用阿里云镜像作为默认源，替代 pypi.org
[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
